BJS2-40653 Стажировка 5/5 (Version 8.0) Fixed all
With all tests.
