+++ src/main/java/faang/school/projectservice/dto/internship/InternshipFilterDto.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.dto.internShip;
+package faang.school.projectservice.dto.internship;
 
 import faang.school.projectservice.model.InternshipStatus;
 import faang.school.projectservice.model.TeamRole;
