+++ src/main/java/faang/school/projectservice/mapper/InternshipMapper.java
@@ -1,10 +1,10 @@
 package faang.school.projectservice.mapper;
 
-import faang.school.projectservice.dto.internShip.InternshipCreatedDto;
-import faang.school.projectservice.dto.internShip.InternshipFilterDto;
-import faang.school.projectservice.dto.internShip.InternshipGetAllDto;
-import faang.school.projectservice.dto.internShip.InternshipGetByIdDto;
-import faang.school.projectservice.dto.internShip.InternshipUpdatedDto;
+import faang.school.projectservice.dto.internship.InternshipCreatedDto;
+import faang.school.projectservice.dto.internship.InternshipFilterDto;
+import faang.school.projectservice.dto.internship.InternshipGetAllDto;
+import faang.school.projectservice.dto.internship.InternshipGetByIdDto;
+import faang.school.projectservice.dto.internship.InternshipUpdatedDto;
 import faang.school.projectservice.model.Internship;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.TeamMember;
