+++ src/test/java/faang/school/projectservice/filter/InternshipStatusFilterTest.java
@@ -1,6 +1,6 @@
 package faang.school.projectservice.filter;
 
-import faang.school.projectservice.dto.internShip.InternshipFilterDto;
+import faang.school.projectservice.dto.internship.InternshipFilterDto;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.TeamRole;
 import org.junit.jupiter.api.Test;
